// src/app/models/vista-estudiante.model.ts
export interface VistaEstudiante {
    idUsuario: number;
    idEstudiante: number;
    activo: boolean;
    correoValido: boolean;
    fechaCreacion: string; 
    nombres: string;
    apellidos: string;
    cedula: string;
    correo: string;
    celular: string;
    facultad: string;
    carrera: string;
}